refactor: address PR review — rename task-model→task-manager, encapsulate model ops
- Rename `task-model.ts` → `task-manager.ts` (and test file) so filename matches class name
- Add `TaskManager.enqueueIssue()` (trackIssue + Task.upsert) and `dequeueIssue()` (untrack + task.delete) to encapsulate coupled model calls that controllers were making separately
- Remove `reconcile()` wrapper from event-router.ts; `wss.ts` wires `reconcile` on the returned ForemanWss directly to `assignIdleWorkers()`; internal call sites use `assignIdleWorkers()` directly
- Clarify `closeIssue()` comment: task.delete() is intentionally a no-op when assigned_at is set, so historically-assigned tasks are always preserved
