+++ CLAUDE.md
@@ -130,7 +130,7 @@ Workers buffer any `task_complete` messages sent during the `hello_sent` state a
 
 A task moves through states: **pending → assigned → pushed → merged/closed → complete** (derived from timestamps and runtime state).
 
-- `Task` (active-record in `task.ts`) owns all DB reads/writes. `TaskManager` (`task-model.ts`) owns only ephemeral state (event queues, branch maps, issue tracking). **There is no in-memory cache — all reads go through `Task` statics.** `initDb(supabase)` from `db-client.ts` must be called once at startup to wire the DB; `TaskManager` subscribes to `Task.events` in its own constructor. Supabase is required to run the foreman; `index.ts` exits with a clear error if credentials are not configured.
+- `Task` (active-record in `task.ts`) owns all DB reads/writes. `TaskManager` (`task-manager.ts`) owns only ephemeral state (event queues, branch maps, issue tracking). **There is no in-memory cache — all reads go through `Task` statics.** `initDb(supabase)` from `db-client.ts` must be called once at startup to wire the DB; `TaskManager` subscribes to `Task.events` in its own constructor. Supabase is required to run the foreman; `index.ts` exits with a clear error if credentials are not configured.
 - The `tasks` table stores `task_id`, `issue_number` (unique), `repo`, `title`, `body`, `labels`, `worker_id`, `assigned_at`, `completed_at`, `issue_closed_at`, `pr_merged_at`. There is no separate `task_assignments` table or stored `status` column.
 - **Status is derived** via the `Task.status` getter based on priorities: `completedAt` → 'complete', `issueClosedAt` → 'closed', `prMergedAt` → 'merged', `workerId` → 'assigned', `prNumber` → 'pushed', `isBlocked()` → 'blocked', else → 'pending'.
 - `Task.upsert()` conflicts on `task_id`. If an issue is re-labeled `brunel:ready` (e.g. after completing), it resets status fields and refreshes `title`, `body`, `labels` — it does **not** ignore duplicates. Returns the `Task` instance.
